requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
django-filter==24.3
Pillow==11.1.0
python-decouple==3.8
drf-spectacular==0.27.2
pyotp==2.9.0
qrcode==8.0
django-ratelimit==4.1.0
